feat: remove redunctunt code
