+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "tidebit-defi",
-  "version": "0.9.0+83",
+  "version": "0.9.0+84",
   "private": true,
   "type": "module",
   "scripts": {
@@ -35,7 +35,6 @@
     "@prisma/adapter-pg": "^7.2.0",
     "@prisma/client": "^7.2.0",
     "apexcharts": "^5.3.6",
-    "cbor-js": "^0.1.0",
     "class-variance-authority": "^0.7.1",
     "clsx": "^2.1.1",
     "jose": "^6.1.3",
@@ -55,7 +54,6 @@
     "@nomicfoundation/hardhat-ignition": "^3.0.6",
     "@nomicfoundation/hardhat-toolbox-viem": "^5.0.1",
     "@tailwindcss/postcss": "^4",
-    "@types/cbor-js": "^0.1.1",
     "@types/chai": "^4.3.20",
     "@types/jest": "^30.0.0",
     "@types/mocha": "^10.0.10",
