+++ src/components/login/register_modal.tsx
@@ -8,99 +8,11 @@ import { RxCross2 } from 'react-icons/rx';
 import { useGlobalCtx } from '@/contexts/global_context';
 import { useModalCtx } from '@/contexts/modal_context';
 import { Button } from '@/components/common/button';
-import { encodeFunctionData, encodeAbiParameters, parseAbiParameters, type Hex } from 'viem';
+import { encodeFunctionData, type Hex } from 'viem';
 import { publicClient } from '@/lib/viem';
 import { CONTRACT_ADDRESSES, ABIS } from '@/config/contracts';
-import { fido2ClientService, parsePasskey, sendUserOpToBundler } from '@/lib/auth/fido2-client';
-import { AuthenticationJSON } from '@passwordless-id/webauthn/dist/esm/types';
-import { base64ToHex, hexToBase64Url } from '@/lib/auth/passkey-encoding';
-
-// Info: (20251226 - Tzuhan) --- [核心] 打包簽名給合約 ---
-function encodeWebAuthnSignature(
-  authentication: AuthenticationJSON,
-  pubKeyX: bigint,
-  pubKeyY: bigint
-): Hex {
-  const { authenticatorData, clientDataJSON, signature } = authentication.response;
-
-  // Info: (20251226 - Tzuhan) 1. 解析 DER 簽名 (提取 r, s)
-  const sigHex = base64ToHex(signature);
-
-  let p = 0;
-  // Info: (20251226 - Tzuhan) Check sequence header (0x30)
-  if (sigHex.substring(p, p + 2) !== '30') throw new Error('Invalid DER signature header');
-  p += 2;
-
-  // Info: (20251226 - Tzuhan) Read total length (並正確跳過)
-  const lenByte = parseInt(sigHex.substring(p, p + 2), 16);
-  p += 2;
-  if (lenByte & 0x80) {
-    // Info: (20251226 - Tzuhan) 如果長度大於 127，會用多個 bytes 表示
-    const lenBytesCount = lenByte & 0x7f;
-    p += lenBytesCount * 2;
-  }
-
-  // Info: (20251226 - Tzuhan) Check Integer Tag for r (0x02)
-  if (sigHex.substring(p, p + 2) !== '02') throw new Error('Invalid DER r tag');
-  p += 2;
-
-  // Info: (20251226 - Tzuhan) Read r length
-  const rLen = parseInt(sigHex.substring(p, p + 2), 16);
-  p += 2;
-
-  // Info: (20251226 - Tzuhan) Read r value
-  const rHex = sigHex.substring(p, p + rLen * 2);
-  const r = BigInt('0x' + rHex);
-  p += rLen * 2;
-
-  // Info: (20251226 - Tzuhan) Check Integer Tag for s (0x02)
-  if (sigHex.substring(p, p + 2) !== '02') throw new Error('Invalid DER s tag');
-  p += 2;
-
-  // Info: (20251226 - Tzuhan) Read s length
-  const sLen = parseInt(sigHex.substring(p, p + 2), 16);
-  p += 2;
-
-  // Info: (20251226 - Tzuhan) Read s value
-  const sHex = sigHex.substring(p, p + sLen * 2);
-  const s = BigInt('0x' + sHex);
-  p += sLen * 2;
-
-  // Info: (20251226 - Tzuhan) 2. 在 ClientDataJSON 中定位 challenge 和 type 的索引
-  let clientDataBase64 = clientDataJSON.replace(/-/g, '+').replace(/_/g, '/');
-  while (clientDataBase64.length % 4) {
-    clientDataBase64 += '=';
-  }
-  const clientDataStr = atob(clientDataBase64);
-
-  const challengePos = clientDataStr.indexOf('"challenge"');
-  // Info: (20251226 - Tzuhan) 尋找值的起始位置 (冒號後的第一個引號後)
-  const challengeValStart = clientDataStr.indexOf('"', challengePos + 11) + 1;
-
-  const typePos = clientDataStr.indexOf('"type"');
-  const typeValStart = clientDataStr.indexOf('"', typePos + 6) + 1;
-
-  // Info: (20251226 - Tzuhan) 3. 使用 viem 編碼為合約結構
-  const encoded = encodeAbiParameters(
-    parseAbiParameters(
-      '(bytes authenticatorData, bytes clientDataJSON, uint256 challengeLocation, uint256 responseTypeLocation, uint256 r, uint256 s, uint256 pubKeyX, uint256 pubKeyY)'
-    ),
-    [
-      {
-        authenticatorData: `0x${base64ToHex(authenticatorData)}`,
-        clientDataJSON: `0x${base64ToHex(clientDataJSON)}`,
-        challengeLocation: BigInt(challengeValStart),
-        responseTypeLocation: BigInt(typeValStart),
-        r,
-        s,
-        pubKeyX,
-        pubKeyY,
-      },
-    ]
-  );
-
-  return encoded;
-}
+import { fido2ClientService, parsePasskey, sendUserOpToBundler } from '@/lib/auth/fido2_client';
+import { encodeWebAuthnSignature, hexToBase64Url } from '@/lib/auth/passkey_encoding';
 
 const RegisterModal: React.FC = () => {
   const [inputValue, setInputValue] = useState<string>('');
