+++ src/lib/auth/fido2_client.ts
