+++ src/lib/auth/fido2_server.ts
