+++ src/lib/hooks/use_passkey_sign.ts
@@ -2,12 +2,12 @@ import { useState } from 'react';
 import { type Hex, encodeAbiParameters, parseAbiParameters } from 'viem';
 import { publicClient } from '@/lib/viem';
 import { CONTRACT_ADDRESSES, ABIS } from '@/config/contracts';
-import { fido2ClientService, sendUserOpToBundler } from '@/lib/auth/fido2-client';
+import { fido2ClientService, sendUserOpToBundler } from '@/lib/auth/fido2_client';
 import {
   hexToBase64Url,
   encodeWebAuthnSignature,
   getWebAuthnSignatureStruct,
-} from '@/lib/auth/passkey-encoding';
+} from '@/lib/auth/passkey_encoding';
 
 export interface IPartialUserOp {
   sender: Hex;
